final fix

+++ createsale.php
@@ -59,7 +59,8 @@ if(Input::exists()){
                 'sale_description' => Input::get('description'),
                 'discount' => Input::get('discount'),
                 'end_date' => date('Y-m-d', strtotime(Input::get('end_date'))),
-                'start_date' => date('Y-m-d', strtotime(Input::get('start_date')))
+                'start_date' => date('Y-m-d', strtotime(Input::get('start_date'))),
+                'client_id' => $user->data()->client_id
             ));
             $isFound = $sale->findSale(array('sale_name', '=', Input::get('sale_name')));
             if(Input::get('good')){
@@ -77,7 +78,8 @@ if(Input::exists()){
             }
 
             Session::flash('createS', 'New sale has been added!');
-            //Redirect::to('index.php');
+
+            Redirect::to('sale.php');
 
         } catch (Exception $e) {
             die($e->getMessage()); //or you can redirect to other page.
@@ -93,6 +95,7 @@ if(Input::exists()){
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <link rel="stylesheet" href="css/bootstrap.min.css">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
     <title>Wecrue - Add Sale</title>
 </head>
 <body>
@@ -191,7 +194,7 @@ if(Input::exists()){
     <h2 class="mb-4">Create sale form</h2>
     <?php
     if(Session::exists('createS')) {
-        echo '<p class="text-success">' . Session::flash('createS') . '</p>';
+        //echo '<p id="created" class="text-success">' . Session::flash('createS') . '</p>';
     }
 
     if($validate->errors()) {
@@ -200,29 +203,33 @@ if(Input::exists()){
         }
     }
     ?>
-    <form action="" method="post">
+    <form action="" method="post" id="content">
         <fieldset class="form-group">
             <legend></legend>
             <div class="form-group col-md-6">
                 <label class="form-control-label" for="sale_name"><span class="text-danger">*</span> Name</label>
-                <input class="form-control" type="text" name="sale_name" id="sale_name" placeholder="Enter name for sale" value="" />
+                <input required maxlength="25" class="form-control" type="text" name="sale_name" id="sale_name" placeholder="Enter name for sale" value="" />
+                <p style="color:red;" id="nameErr"></p>
             </div>
             <div class="form-group col-md-6">
                 <label class="form-control-label" for="description"><span class="text-danger">*</span> Description</label>
-                <textarea class="form-control" rows="3" name="description" id="description" placeholder="Enter description"><?php echo escape(Input::get('client_info'))?></textarea>
+                <textarea required maxlength="130" class="form-control" rows="3" name="description" id="description" placeholder="Enter description"><?php echo escape(Input::get('client_info'))?></textarea>
             </div>
             <div class="form-group form-inline">
                 <label class="form-control-label mr-2 ml-3" for="discount"><span class="text-danger">*</span>&nbspDiscount</label>
-                <input class="form-control" type="number" min="0.01" step="0.01" name="discount" id="discount" style="width: 90px;" value="" />
+                <input required class="form-control" type="number" min="0.01" max="99.99" step="0.01" name="discount" id="discount" style="width: 90px;" value="10" />
                 <span class="input-group-addon">%</span>
+                <p style="color:red;" id="discountErr"></p>
             </div>
             <div class="form-group col-md-6" >
                 <label class="form-control-label" for="start_date"><span class="text-danger">*</span> Start date</label>
-                <input class="form-control" type="date" name="start_date" id="start_date" placeholder="Format: YYYY-MM-DD" />
+                <input required class="form-control" type="date" name="start_date" id="start_date" placeholder="Format: YYYY-MM-DD" />
+                <p style="color:red;" id="startErr"></p>
             </div>
             <div class="form-group col-md-6">
                 <label class="form-control-label" for="end_date"><span class="text-danger">*</span> End date</label>
-                <input class="form-control" type="date" name="end_date" id="end_date" placeholder="Format: YYYY-MM-DD" />
+                <input required class="form-control" type="date" name="end_date" id="end_date" placeholder="Format: YYYY-MM-DD" />
+                <p style="color:red;" id="endErr"></p>
             </div>
             <div class="form-control-label">
                 <p class="text-info"><span class="text-danger">*</span>Select product or category</p>
@@ -237,7 +244,10 @@ if(Input::exists()){
                         $categoryItems = $category->data();
 
                         foreach ($categoryItems as $categoryItem) {
-                            echo '<option value="' . $categoryItem->category_id . '">' . $categoryItem->category_name . '</option>';
+                            $good->getGood(array('category_id', '=', $categoryItem->category_id));
+                            if($good->exists()){
+                                echo '<option value="' . $categoryItem->category_id . '">' . $categoryItem->category_name . '</option>';
+                            }
                         }
                     }
                     ?>
@@ -249,6 +259,7 @@ if(Input::exists()){
                     <option value="">Select good</option>
                     <?php
                     //$category->getCategory($user->data()->client_id);
+                    $hasGood = false;
                     if($category->exists()) {
                         $categoryItems = $category->data();
 
@@ -258,11 +269,16 @@ if(Input::exists()){
                                 $goodItems = $good->data();
 
                                 foreach ($goodItems as $goodItem){
-                                    if (!$goodItem->sale_id) echo '<option value="' . $goodItem->good_id . '">' . $goodItem->good_name . '</option>';
+
+                                    if (!$goodItem->sale_id){
+                                        echo '<option value="' . $goodItem->good_id . '">' . $goodItem->good_name . '</option>';
+                                        $hasGood = true;
+                                    }
                                 }
                             }
                         }
                     }
+
                     ?>
                 </select>
             </div>
@@ -293,7 +309,7 @@ if(Input::exists()){
         </fieldset>
         <div class="form-group">
             <input type="hidden" name="token" value="<?php echo Token::generate(); ?>">
-            <input class="btn btn-primary" type="submit" value="Submit" />
+            <input id="subm" class="btn btn-primary" type="submit" value="Submit" />
         </div>
     </form>
 </div>
@@ -302,6 +318,99 @@ if(Input::exists()){
 <script src="js/jquery-3.1.1.slim.min.js"></script>
 <script src="js/tether.min.js"></script>
 <script src="js/bootstrap.min.js"></script>
+<script>
+    var subm = function() {
+        if ( $("#nameErr").text() == "" && $("#startErr").text() == "" && $("#endErr").text() == ""){
+            $("#subm").removeAttr("disabled");
+        }else{
+            $("#subm").attr("disabled","disabled");
+        }
+    }
+
+    $("#discount").blur(function(){
+        var discount = $("#discount").val();
+        discount = parseFloat(discount).toFixed(2);
+        if(discount < 0 || discount >= 100){
+            $("#discount").val("10");
+            $("#discountErr").text("* Discount must be between 0 ~ 99.99%");
+        }else{
+            $("#discount").val(discount);
+            $("#discountErr").text("");
+        }
+        subm();
+    });
+
+    var check = function(){
+        if($("#start_date").val() && $("#start_date").val() == $("#end_date").val()){
+            $("#endErr").text("* start date and end date cannot be same");
+        }else{
+            $("#endErr").text("");
+        }
+
+
+        if($("#start_date").val() != "" ){
+            var today = new Date();
+            var sDate=new Date();
+            var s = $("#start_date").val().split("-");
+            sDate.setFullYear(s[0],s[1]-1,s[2]);
+            if(sDate <= today){
+                $("#startErr").text("* start date must be a future date");
+            }else{
+                $("#startErr").text("");
+            }
+        }
+
+        if($("#start_date").val() != "" && $("#end_date").val()){
+            var today = new Date();
+            var sDate=new Date();
+            var s = $("#start_date").val().split("-");
+            sDate.setFullYear(s[0],s[1]-1,s[2]);
+
+            var eDate=new Date();
+            var e = $("#end_date").val().split("-");
+            eDate.setFullYear(e[0],e[1]-1,e[2]);
+
+            if(sDate >= eDate){
+                $("#endErr").text("* end date must be greater than start date");
+            }else{
+                $("#endErr").text("");
+            }
+        }
+        subm();
+    };
+
+    $("#start_date").blur(function(){
+        check();
+    });
+
+    $("#end_date").blur(function(){
+        check();
+    });
+
+    $("#sale_name").blur(function(){
+        var name = $("#sale_name").val();
+        $.ajax({
+          url: "tools/checkSale.php",
+          data: {
+            name: name
+          },
+          dataType: "text",
+          method: "GET",
+          success: function(data) {
+            $("#nameErr").text(data);
+            subm();
+          }
+        });
+    });
+        <?php
+
+    if(!$hasGood ){
+        $str = "<p>You don't have any product to create sale. <a href='create-good.php'>Click here to create one.</a></p>";
+        echo '$("#content").html("'.$str.'")';
+    }
+        ?>
+
+</script>
 
 </body>
 </html>
